requirements.txt
python-binance>=1.0.17
numpy>=1.23
pandas>=1.5